Addition of placeholder battle pane
Entering a tile with a monster will now trigger entry into battle mode.
Battles with enemy NPCs will occur in battle mode. There will be action
choices in a battle pane, which is currently represented by a
placeholder window. Messages will be displayed in either the
messages/debug window or the battle window.